lab3/project/auth_service/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.3
pika==1.2.0
